refactor: refactor menu package
